Title: 고인 물, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241115/고인 물/standing-water.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int n, m;
 priority_queue<int, vector<int>, greater<int>> pq;
-int answer = 0;
+unsigned long answer = 0;
 int board[505];
 
 int main()
